// src/app/core/models/task.model.ts
export interface Task {
    id: string;
    title: string;
    description: string;
    completed: boolean;
    createdAt: string;
    userId: string;
}

export interface CreateTaskInput {
    title: string;
    description?: string;
}

export interface UpdateTaskInput {
    title?: string;
    description?: string;
    completed?: boolean;
}